Refactored the Transaction and TransactionBuilder classes
The TransactionBuilder class now queries the transaction database to ensure it has the most up-to-date information for the Doi objects it was provided.
The updated Doi objects are then used by TransactionBuilder to create the output content label stored by Transaction.log() to the local filesystem.
The output content can also now be fetched from a Transaction object via the output_content property.