Updated DOI API spec to version 0.2
This version updates several endpoints to not include {lidvid} as part of the path, but rather as a query parameter.
This resolves the issue with forward slashes being present in some PDS3 identifiers getting used as lidvids.
The update also replaces all references to "lid" or "lidvid" with to the more generic "identifier".